+++ tests/test.js
@@ -260,8 +260,17 @@
         call: function (message) {
           var hash = algorithm.update(message);
           hash.hex();
-          hash.update(message);
+          hash.finalize();
           return hash.hex();
+        }
+      },
+      {
+        name: 'clone',
+        call: function (message) {
+          var hash = algorithm.update(message);
+          var hash2 = hash.clone();
+          hash.update('any');
+          return hash2.hex();
         }
       }
     ];
@@ -314,6 +323,16 @@
               algorithm(testCase);
             }).to.throwError(/input is invalid type/);
           });
+        });
+      });
+
+      context('when update after finalize', function () {
+        it('should throw error', function () {
+          expect(function () {
+            var hash = algorithm.update('any');
+            hash.hex();
+            hash.update('any');
+          }).to.throwError(/finalize already called/);
         });
       });